OO-6421: Media support for static text launcher

+++ src/main/java/org/olat/modules/catalog/manager/CatalogV2ServiceImpl.java
@@ -262,7 +262,14 @@ public class CatalogV2ServiceImpl implements CatalogV2Service {
 
 	@Override
 	public void deleteCatalogLauncher(CatalogLauncherRef catalogLauncher) {
-		catalogLauncherDao.delete(catalogLauncher);
+		CatalogLauncher reloadedLancher = getCatalogLauncher(catalogLauncher);
+		if (reloadedLancher != null) {
+			CatalogLauncherHandler handler = getCatalogLauncherHandler(reloadedLancher.getType());
+			if (handler != null) {
+				handler.deleteLauncherData(reloadedLancher);
+			}
+			catalogLauncherDao.delete(catalogLauncher);
+		}
 	}
 	
 	@Override
